requirements.txt
functions-framework>=3.5.0
pydantic>=1.10,<2.0
google-cloud-bigquery>=3.11.0
pytest-mock
pytest